remove dependency from httpclient to h1/h2 implementations
